Change assertion runner to catch AssertionError
